refactor interactWithDM procedure into its own class
- refactor interaction as to avoid code duplication
- tweak NSGAII so its test remains working
